Moved the permission request to the Voice Mode screen instead of on navigation click
