+++ chat/feat-chat/src/main/java/com/seravian/feat_chat/data/AudioStreamer.kt
@@ -85,7 +85,7 @@ class AudioStreamer(
 
                     val currentTime = System.currentTimeMillis()
 
-                    if (rms > 500) {
+                    if (rms > 250) {
                         lastVoiceTime = currentTime
                         voiceDetected = true
                     }
