+++ chat/feat-chat/src/main/java/com/seravian/feat_chat/presentation/viewModel/ChatViewModel.kt
@@ -265,13 +265,14 @@ class ChatViewModel(
     }
 
     private fun stopStreaming() {
+        _voiceState.update {
+            it.copy(
+                isStreamingVoice = false
+            )
+        }
+
         if (::audioStreamer.isInitialized) {
             audioStreamer.stop()
-            _voiceState.update {
-                it.copy(
-                    isStreamingVoice = false
-                )
-            }
         }
     }
